Allow debugging Workbox from localhost
